+++ build.gradle
@@ -52,21 +52,20 @@ jar {
 
 // configure the maven publication
 publishing {
-	publications {
-		mavenJava(MavenPublication) {
-			// add all the jars that should be included when publishing to maven
-			artifact(remapJar) {
-				builtBy remapJar
-			}
-			artifact(sourcesJar) {
-				builtBy remapSourcesJar
+	repositories {
+		maven {
+			name = "GitHubPackages"
+			url = uri("https://maven.pkg.github.com/hackcoder1000/fabricshieldlib")
+			credentials {
+				username = property('gpr.username')
+				password = property('gpr.token')
 			}
 		}
 	}
-
-	// select the repositories you want to publish to
-	repositories {
-		// uncomment to publish to the local maven
-		// mavenLocal()
+	publications {
+		gpr(MavenPublication) {
+			from(components.java)
+			artifact sourcesJar
+		}
 	}
 }
