+++ gradle.properties
@@ -3,13 +3,13 @@ org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties
 # check these on https://fabricmc.net/use
-minecraft_version=1.17
-yarn_mappings=1.17+build.13
+minecraft_version=1.17.1
+yarn_mappings=1.17.1+build.32
 loader_version=0.11.6
 
 # Mod Properties
 mod_version = 1.3.7
-maven_group = me.crimsondawn45
+maven_group = net.legiblesleet827
 archives_base_name = FabricShieldLib
 
 # Dependencies
