--- jitpack.yml
@@ -1,5 +1,0 @@
-before_install:
-  - export SDKMAN_DIR="/home/jitpack/sdkman" && curl -s "https://get.sdkman.io" | bash
-  - source "/home/jitpack/sdkman/bin/sdkman-init.sh"
-  - sdk install java 16.0.1.hs-adpt
-  - sdk use java 16.0.1.hs-adpt
