+++ settings.gradle
@@ -8,3 +8,4 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
+rootProject.name = "fabricshieldlib"
