+++ src/main/java/net/legiblesleet827/fabricshieldlib/lib/event/ShieldEvent.java
@@ -1,4 +1,4 @@
-package me.crimsondawn45.fabricshieldlib.lib.event;
+package net.legiblesleet827.fabricshieldlib.lib.event;
 
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
