+++ src/main/resources/fabricshieldlib.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "me.crimsondawn45.fabricshieldlib.mixin",
+  "package": "net.legiblesleet827.fabricshieldlib.mixin",
   "compatibilityLevel": "JAVA_16",
   "mixins": [
     "PlayerEntityMixin",
